speed bonus is float, not int

+++ src/main/java/team/cqr/cqrepoured/item/IExtendedItemTier.java
@@ -5,7 +5,7 @@ import net.minecraft.item.IItemTier;
 public interface IExtendedItemTier extends IItemTier {
 	
 	public int getFixedAttackDamageBonus();
-	public int getAttackSpeedBonus();
+	public float getAttackSpeedBonus();
 	public double getMovementSpeedBonus();
-
+	
 }
